Minor improvements (#1494)
* Minor improvements

* Update CHANGELOG.md

+++ library/includes/datastore/build_files_integrity.php
@@ -69,7 +69,9 @@ if (is_file(RESTORE_CORRUPT_CONFIRM_FILE)) {
     }
 
     // Delete restore confirm file & build file
-    unlink($buildDownloader->savePath);
+    if (isset($buildDownloader->savePath)) {
+        unlink($buildDownloader->savePath);
+    }
     unlink(RESTORE_CORRUPT_CONFIRM_FILE);
 }
